fix salary .00

+++ lab2/soa-lab2-server1/src/main/java/com/example/soalab2server1/dao/request/WorkerInfo.java
@@ -40,7 +40,7 @@ public class WorkerInfo {
     @PositiveOrZero
     @NotNull
     @Digits(integer = Integer.MAX_VALUE , fraction = 2)
-    private float salary;
+    private double salary;
 
     @JacksonXmlProperty(localName = "startDate")
     @NotNull
